[MP-005] Updated pageSize and maxPageSize from Long to Int to match supported DB and API appropriate typing (limited by Int).

+++ scheduler-quartz/src/main/kotlin/io/scheduler/comparison/quartz/config/properties/StaticOrderJobProperties.kt
@@ -27,10 +27,10 @@ data class StaticOrderJobProperties(
         val cron: String,
         @field:Min(1)
         @field:Max(1000)
-        val pageSize: Long,
+        val pageSize: Int,
         @field:Min(1)
         @field:Max(1000)
-        val maxCountPerExecution: Long,
+        val maxCountPerExecution: Int,
     )
 
     @Validated
@@ -48,10 +48,10 @@ data class StaticOrderJobProperties(
         val jobHandler: String = "dedicatedJobHandler",
         @field:Min(1)
         @field:Max(1000)
-        val pageSize: Long,
+        val pageSize: Int,
         @field:Min(1)
         @field:Max(1000)
-        val maxCountPerExecution: Long,
+        val maxCountPerExecution: Int,
     )
 
 }
