+++ scheduler-quartz/src/main/kotlin/io/scheduler/comparison/quartz/jobs/DedicatedOrderJob.kt
@@ -63,8 +63,8 @@ class DedicatedOrderJob : Job {
         = DedicatedOrderJobMetadata(
         jobName = jobDataMap.getString(DedicatedOrderJobParams.JOB_NAME.value),
         jobCron = jobDataMap.getString(DedicatedOrderJobParams.JOB_CRON.value),
-        maxCountPerExecution = jobDataMap.getLong(DedicatedOrderJobParams.MAX_COUNT_PER_EXECUTION.value),
-        pageSize = jobDataMap.getLong(DedicatedOrderJobParams.PAGE_SIZE.value),
+        maxCountPerExecution = jobDataMap.getInt(DedicatedOrderJobParams.MAX_COUNT_PER_EXECUTION.value),
+        pageSize = jobDataMap.getInt(DedicatedOrderJobParams.PAGE_SIZE.value),
     )
 
 }
