+++ scheduler-quartz/src/main/kotlin/io/scheduler/comparison/quartz/jobs/state/PaginatedJobMetadata.kt
@@ -2,7 +2,7 @@ package io.scheduler.comparison.quartz.jobs.state
 
 interface PaginatedJobMetadata : JobMetadata {
 
-    val pageSize: Long
-    val maxCountPerExecution: Long
+    val pageSize: Int
+    val maxCountPerExecution: Int
 
 }
